Finra License Search -2 Stable Testing
All aspects of the BDG Enhancement should now be operational and working
in code.

Will test on a live list today, 6/22/2016, to ensure the program works
appropriately.

Prior to testing data in the test BDG will be capture so that we can
revert, if necessary, to the old information in SFDC.