trainHeadersModule: Enhanced Matching
in the count vectorizer, switched the code from analyzer=word to
analyzer=char. This should help to increase the match rate.

Essentially, the anlyzer=word meant that the vectors, or unique
features, of the headers were being created lexically (looking for
similarity ie. Address 1 and Address 2 return the prediction
MailingStreet1) so the error rate was high.
With the update, it now looks at individual characters of the words ie
[a,d,d,r,e,s,s, ,1] and [a,d,d,r,e,s,s, 2]. It now understands these as
two vastly different