incorporating list object stats integration
resolves #81 resolves #84  resolves #83  resolves #74

in the processes and list processing modules we now have the ability to
interact with a list record if 1) we're successfully able to process the
list or 2) if the file is of a bad extension type fixes #81

in the list processing module, where keys are being cleared after each
list, instead of mutating (deleting) dictionary keys - we now just set
the values to None so that processing can continue if there's more than
1 (one) list in the queue. fixes #74

in the email_reader module, emails are no longer sent to the individual
that uploaded the list when the list first starts processing. instead,
they will only received a completion, or unable to process, email. fixes
#83

in the email_reader module, the MailBoxReader class now has a
close_mailbox method to shut down the mailbox instance. fixes #84