enhance SFDC upload processing and integrade processing
fixes #91 resolves #91

SFPlatform upload_attachements method is now working.

introduced a way to dynamically ensure that every file gets uploaded
(regardless of the length of the path & file name; max is 120 chars) by
moving it to a temporary folder and uploading from there. after all
attachments in the request are uploaded, the temporary folder (and it's
contents) are deleted.

integrated the enhanced SFPlatform functionality into the processes.py
module.