adding diagnostics method to the ML api
this allows a user to manually, or on a schedule, find the best module
for our header prediction problem based on the 7-fold cross validation.

additionally, add libraries to the __init__.py directory in the main
folder to bring classes to the top-level for interactions like a normal
package.

provide examples of how to interact, debug , test specifc classes