gen_helper/determine_ext update
There's currently no need (I'm aware of) to check if an extention is in
a list of values. There are other checks, in the appropriate modules, to
check for _acceptable_file_types. This check seems redudant and will
lead to the extention of the list in the function any time there is
something new that causes an error.

In an effort to streamline this to not need patch fixes, I'm removing
the check from the code base.