initial commit to address issues and fill gaps in processing
#138
a fix is included in this commit to address the tree structure of list
files.

#137
a fix is included to build the data, and then upload to salesforce, for
the last list upload

#134
an inital build of an 'auto_maintain' function was added to server as
the framework of a more robust function

#139
update legacy.* files refer to appropriate libraries in pandas.

#140
re-configure log_helper.py

#141
begin addressing/replacing all print statements with appropriate
log.<action> method

general naming refactoring/process refinement throughout the codebase